+++ applications/GetGFSAnalysisFromRDA.csh
@@ -48,6 +48,17 @@ cd ${WorkDir}
 
 # ================================================================================================
 
+if ( -e GETSUCCESS ) then
+  echo "$0 (INFO): GETSUCCESS file already exists, exiting with success"
+  echo "$0 (INFO): if regenerating the output files is desired, delete GETSUCCESS"
+
+  date
+
+  exit 0
+endif
+
+# ================================================================================================
+
 set linkWPS = link_grib.csh
 ln -sfv ${WPSBuildDir}/${linkWPS} .
 rm -rf GRIBFILE.*
@@ -74,5 +85,7 @@ if ( ! -e "GRIBFILE.AAA") then
 endif
 
 date
+
+touch GETSUCCESS
 
 exit 0
